refactor(cli): own native compiler execution
