+++ crates/oleafly-cli/Cargo.toml
@@ -15,8 +15,29 @@ path = "src/main.rs"
 clap = { version = "4.6", features = ["derive"] }
 notify = "8.2"
 oleafly-core = { path = "../oleafly-core" }
+serde = { version = "1", features = ["derive"] }
 serde_json = "1"
-tokio = { version = "1", features = ["macros", "rt-multi-thread", "signal", "sync", "time"] }
+sha2 = "0.10"
+tokio = { version = "1", features = [
+  "io-util",
+  "macros",
+  "process",
+  "rt-multi-thread",
+  "signal",
+  "sync",
+  "time",
+] }
+
+[target.'cfg(unix)'.dependencies]
+libc = "0.2"
+
+[target.'cfg(windows)'.dependencies]
+windows-sys = { version = "0.61", features = [
+  "Win32_Foundation",
+  "Win32_System_Diagnostics_ToolHelp",
+  "Win32_System_JobObjects",
+  "Win32_System_Threading",
+] }
 
 [dev-dependencies]
 tempfile = "3.27"
