+++ crates/oleafly-core/Cargo.toml
@@ -10,20 +10,6 @@ rust-version = "1.77"
 [dependencies]
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
-sha2 = "0.10"
-tokio = { version = "1", features = ["io-util", "process", "rt", "sync", "time"] }
-
-[target.'cfg(unix)'.dependencies]
-libc = "0.2"
-
-[target.'cfg(windows)'.dependencies]
-windows-sys = { version = "0.61", features = [
-  "Win32_Foundation",
-  "Win32_System_Diagnostics_ToolHelp",
-  "Win32_System_JobObjects",
-  "Win32_System_Threading",
-] }
 
 [dev-dependencies]
 tempfile = "3.27"
-tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
