+++ crates/oleafly-core/src/lib.rs
@@ -1,12 +1,9 @@
 mod build;
 mod error;
 mod manifest;
-mod process;
 mod workspace;
 
-pub use build::{
-    BuildError, BuildOptions, BuildResult, BuildTools, CompilerLog, NativeCompiler, PreparedBuild,
-};
+pub use build::PreparedBuild;
 pub use error::{Error, ErrorKind, Result};
 pub use manifest::{Engine, ExportRecord, ProjectManifest, TexSpec};
 pub use workspace::{DoctorCheck, DoctorReport, DoctorStatus, InitOptions, ProjectInfo, Workspace};
